requirements.txt
streamlit
tensorflow
Pillow

matplotlib
numpy
pandas
